Cleaned up compile warnings.

+++ branches/releases/bacnet-stack-0-8-0/demo/handler/h_awf.c
@@ -130,7 +130,7 @@ void handler_atomic_write_file(
 #if PRINT_ENABLED
                 fprintf(stderr, "AWF: Stream offset %d, %d bytes\n",
                     data.type.stream.fileStartPosition,
-                    octetstring_length(&data.fileData));
+                    (int)octetstring_length(&data.fileData));
 #endif
                 len =
                     awf_ack_encode_apdu(&Handler_Transmit_Buffer[pdu_len],
